Fix to application task exception

+++ rcrpapplications/rcrpapplications.py
@@ -25,41 +25,51 @@ class RCRPApplications(commands.Cog, name='RCRP Applications'):
     def cog_unload(self):
         self.check_pending_applications.cancel()
 
+    async def log(self, message: str):
+        rcrpguild = self.bot.get_guild(rcrpguildid)
+        logchannel = rcrpguild.get_channel(775767985586962462)
+        await logchannel.send(message)
+
     @tasks.loop(seconds=5.0)
     async def check_pending_applications(self):
-        rcrpguild = await self.bot.fetch_guild(rcrpguildid)
-        appchannel: discord.TextChannel = self.bot.get_channel(appchannelid)
-        sql = await aiomysql.connect(**mysqlconfig)
-        cursor: Cursor = await sql.cursor(aiomysql.DictCursor)
-        await cursor.execute("SELECT quizapps.id, maID, masters.Username, masters.EMail, characterName, quizapps.created_at, ipscore FROM quizapps JOIN masters ON quizapps.maID = masters.id WHERE quizapps.state = 0")
+        try:
+            rcrpguild = await self.bot.fetch_guild(rcrpguildid)
+            appchannel: discord.TextChannel = self.bot.get_channel(appchannelid)
+            sql = await aiomysql.connect(**mysqlconfig)
+            cursor: Cursor = await sql.cursor(aiomysql.DictCursor)
+            await cursor.execute("SELECT quizapps.id, maID, masters.Username, masters.EMail, characterName, quizapps.created_at, ipscore FROM quizapps JOIN masters ON quizapps.maID = masters.id WHERE quizapps.state = 0")
 
-        application_ids = []
-        if cursor.rowcount != 0:
-            data = await cursor.fetchall()
-            for row in data:
-                application_ids.append(row['id'])
-                if row['id'] not in self.applications:
-                    embed = discord.Embed(title="Click here to go to the application", url=f"https://redcountyrp.com/admin/applications/{row['id']}", color=0x008080)
-                    embed.set_author(name="RCRP Application", url=f"https://redcountyrp.com/admin/applications/{row['id']}", icon_url=rcrpguild.icon_url)
-                    embed.add_field(name="Username", value=row['Username'], inline=True)
-                    embed.add_field(name="Email", value=row['EMail'], inline=True)
-                    embed.add_field(name="Character ", value=row['characterName'], inline=True)
-                    embed.add_field(name="IP Score", value=row['ipscore'], inline=True)
-                    embed.timestamp = row['created_at']
-                    message: discord.Message = await appchannel.send(embed=embed)
-                    self.applications[row['id']] = message.id
-        else:
-            messages = await appchannel.history().filter(pinned_filter).flatten()
-            await appchannel.delete_messages(messages)
-            self.applications.clear()
+            application_ids = []
+            if cursor.rowcount != 0:
+                data = await cursor.fetchall()
+                for row in data:
+                    application_ids.append(row['id'])
+                    if row['id'] not in self.applications:
+                        embed = discord.Embed(title="Click here to go to the application", url=f"https://redcountyrp.com/admin/applications/{row['id']}", color=0x008080)
+                        embed.set_author(name="RCRP Application", url=f"https://redcountyrp.com/admin/applications/{row['id']}", icon_url=rcrpguild.icon_url)
+                        embed.add_field(name="Username", value=row['Username'], inline=True)
+                        embed.add_field(name="Email", value=row['EMail'], inline=True)
+                        embed.add_field(name="Character ", value=row['characterName'], inline=True)
+                        embed.add_field(name="IP Score", value=row['ipscore'], inline=True)
+                        embed.timestamp = row['created_at']
+                        message: discord.Message = await appchannel.send(embed=embed)
+                        self.applications[row['id']] = message.id
+            else:
+                messages = await appchannel.history().filter(pinned_filter).flatten()
+                if len(messages) != 0:
+                    await appchannel.delete_messages(messages)
+                    self.applications.clear()
 
-        if len(self.applications) != 0:
-            for key in self.applications:
-                if key not in application_ids:
-                    message = await appchannel.fetch_message(self.applications[key])
-                    await message.delete()
+            if len(self.applications) != 0:
+                for key in self.applications:
+                    if key not in application_ids:
+                        message = await appchannel.fetch_message(self.applications[key])
+                        await message.delete()
+                        del self.applications[key]
 
-        await appchannel.edit(name=f'applications-{cursor.rowcount}')
+            await appchannel.edit(name=f'applications-{cursor.rowcount}')
+        except Exception as e:
+            await self.log(f'{e}')
 
     @check_pending_applications.before_loop
     async def before_check_pending_applications(self):
